requirements/dev.txt
-r ./base.txt

factory-boy==3.3.0
Faker==20.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-factoryboy==2.6.0
django-dotenv==1.4.2
